% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompressiveNMF_MultiStudy.R
\name{CompressiveNMF_MultiStudy}
\alias{CompressiveNMF_MultiStudy}
\title{Maximum-a-posteriori for the CompressiveNMF algorithm}
\usage{
CompressiveNMF_MultiStudy(
  X,
  cohorts_num,
  K = 20,
  S = NULL,
  alpha = 1,
  a = 1,
  epsilon = 0.001,
  cutoff = 5 * epsilon,
  hierarchy = FALSE,
  compressive = TRUE,
  tol = 1e-07,
  a0 = epsilon * 1,
  b0 = 1,
  maxiter = 1e+06
)
}
\arguments{
\item{X}{Matrix of mutations (rows are mutations, columns are patients)}

\item{K}{Upper bound to the number of signatures}

\item{S}{Matrix of informative prior for the signatures}

\item{alpha}{Parameter for the Dirichlet prior in the signature}

\item{a}{Parameter for the gamma prior over the loadings}

\item{epsilon}{Parameter controlling the mean of the Compressive hyperprior}

\item{cutoff}{Cutoff for excluded signatures}

\item{tol}{Tolerance for convergence. Set to \code{1e-6}}

\item{a0}{Shape parameter for the inverse gamma compressive hyperprior. Default is \code{a * ncol(X) + 1}}

\item{b0}{Rate parameter for the inverse gamma compressive hyperprior. Default is \code{a * epsilon * ncol(X)}}

\item{maxiter}{Maximum number of iterations for the algorithm. Set to  \code{1e6}}
}
\description{
Maximum-a-posteriori for the CompressiveNMF algorithm
}
