% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompressiveNMF_random_a.R
\name{CompressiveNMF_randa}
\alias{CompressiveNMF_randa}
\title{Function for our compressive NMF methods}
\usage{
CompressiveNMF_randa(
  X,
  K = 20,
  nsamples = 1000,
  burnin = 2000,
  alpha = 0.5,
  a = 1,
  epsilon = 0.001,
  nchains = 1,
  ncores = 1,
  a0 = a * ncol(X) + 1,
  b0 = a * epsilon * ncol(X),
  random_a = FALSE,
  c0 = 1,
  d0 = 1,
  S = NULL,
  cutoff_excluded = 5 * epsilon,
  use_cosmic = FALSE,
  swap_prior = TRUE,
  betah = 100,
  betah_optimal = TRUE,
  verbose = TRUE,
  seed = 10
)
}
\arguments{
\item{X}{The input data matrix.}

\item{K}{Number of signatures.}

\item{nsamples}{Number of samples in each MCMC chain.}

\item{burnin}{Number of burn-in iterations in each MCMC chain.}

\item{alpha}{Parameter for the Dirichlet prior in the signature}

\item{a}{Parameter for the gamma prior over the loadings}

\item{epsilon}{Parameter controlling the mean of the Compressive hyperprior}

\item{nchains}{Number of chains to run in parallel.}

\item{ncores}{Number of CPU cores to use.}

\item{a0}{Shape parameter for the inverse gamma compressive hyperprior. Default is \code{a * ncol(X) + 1}}

\item{b0}{Rate parameter for the inverse gamma compressive hyperprior. Default is \code{a * epsilon * ncol(X)}}

\item{random_a}{Whether a should be treated as a random quantity}

\item{c0}{Prior for a}

\item{d0}{Prior for a}

\item{S}{Matrix of informative prior for the signatures}

\item{cutoff_excluded}{Cutoff for excluded signatures}

\item{use_cosmic}{Logical, indicating whether to use Cosmic data.}

\item{swap_prior}{Logical, indicating whether to swap the informative prior at 2/3 of the burnin phase.}

\item{betah}{Parameter controlling the strenght of the prior for the informative priors in S. Length has to be equal to the number of columns of \code{S}}

\item{betah_optimal}{Logical, indicating whether to use an optimal betah. See \code{src/tune_betah.cpp}}

\item{verbose}{Logical, indicating whether to display verbose output.}

\item{seed}{Seed for reproducibility.}

\item{progressbar}{Logical, indicating whether to show a progress bar.}
}
\value{
Returns a list of results from CompressiveNMF.
}
\description{
Function for our compressive NMF methods
}
