% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signatures.R
\name{plot.CompressiveNMF}
\alias{plot.CompressiveNMF}
\title{Plotting function for CompressiveNMF}
\usage{
\method{plot}{CompressiveNMF}(object, type = "signatures", ...)
}
\arguments{
\item{object}{An object of class \code{CompressiveNMF} description}
}
\description{
Plotting function for CompressiveNMF
}
