% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signatures.R
\name{plot.DBS.signature}
\alias{plot.DBS.signature}
\title{Plot Double-based-substitution signatures}
\usage{
\method{plot}{DBS.signature}(
  signatures,
  lowCI = NULL,
  highCI = NULL,
  palette = c("#03BDEE", "#0366CD", "#A3CE62", "#076501", "#FD9798", "#E42925",
    "#FEB064", "#FD8004", "#CB99FC", "#4B0C9B")
)
}
\arguments{
\item{signatures}{Matrix of SBS signatures}

\item{lowCI}{Matrix of lower credible interval}

\item{highCI}{Matrix of lower credible interval}
}
\description{
Plot Double-based-substitution signatures
}
