% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signatures.R
\name{plot.ID.signature}
\alias{plot.ID.signature}
\title{Plot Indel signatures}
\usage{
\method{plot}{ID.signature}(
  signatures,
  lowCI = NULL,
  highCI = NULL,
  palette = c("#FEBE6E", "#FD8004", "#A3CE62", "#076501", "#FBCAB4", "#FC886F",
    "#EE4635", "#C11717", "#D1E1F2", "#94C2E0", "#4C95C8", "#1862AB", "#E4E3E8",
    "#B5B7D8", "#8584BD", "#614199")
)
}
\arguments{
\item{signatures}{Matrix of SBS signatures}

\item{lowCI}{Matrix of lower credible interval}

\item{highCI}{Matrix of lower credible interval}
}
\description{
Plot Indel signatures
}
